Improve code readabillity by reformatting and rewriting comments
The default GNOME builder template for GTK C applications contains a
lot of useless comments (for a real project, it's great for an example
template), removing them improves readabillity of the code. Also I
reverted a few of the formatting decisions done by clang-tidy, not sure
how to automate this yet.